Created separate package for @ContextConfiguration class usage question
